pyproject.toml
[project]
name = "intranetFA"
version = "0.1.0"
description = "Projet de création d'un intranet"
authors = [
    {name = "Ewen Vuichard",email = "stage.si@fougeres-agglo.bzh"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-summernote (>=0.8.20.0,<0.9.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "pillow-avif-plugin (>=1.5.2,<2.0.0)",
    "ldap3 (>=2.9.1,<3.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "coverage (>=7.11.0,<8.0.0)",
    "djlint (>=1.36.4,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "intranetFA"}]

[tool.djlint]
indent = 2
profile = "django"
max_line_length = 120
format_attribute_template_tags = true
ignore = "H006,H030,H031"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
